member and admin
会员和管理员完成，下一步完成管理员对学科专业的管理

+++ WebRoot/app/model/Project.js
@@ -8,7 +8,7 @@ Ext.define("Project", {
 		{name : 'delegationUnit',type:'string',sortable:true},
 		{name : 'projectLevel',type:'int',sortable:true},
 		{name : 'completeStatus',type:'int',sortable:true},
-		{name : 'allocation',type:'int',sortable:true},
-		{name : 'selffinance',type:'int',sortable:true}
+		{name : 'allocation',type:'float',sortable:true},
+		{name : 'selffinance',type:'float',sortable:true}
 	]
 });
